save progress
trying out a bunch of things
mostly working as I expect.

one thing that is bugging me:
for already signed in user, we they open the app, why doesn't it automatically log them in?
can fix it outside of firebaseui component, by just interfacing with the auth service
that can cause data races though. need to be more thoughtful in this.
and experiment!